+++ src/components/modal/index.ts
@@ -1,4 +1,2 @@
-export * from "./modalContext";
-export * from "./ModalProvider";
 export * from "./modal";
 export * from "./useModalHook";
