--- src/components/modal/modalContext.tsx
@@ -1,13 +1,0 @@
-import { createContext } from "react";
-
-
-type ModalContextType = {
-  isOpen: boolean;
-  currentModal: string | null;
-  openModal: (modalId: string) => void;
-  closeModal: () => void;
-};
-
-export const ModalContext = createContext<ModalContextType | undefined>(
-  undefined,
-);
